Move FilterBuilder from Helper to Model\Filter
